Iterate all_items for get_file_attachments
Otherwise we will support 25 images only because
you cannot remove image 26 or higher from that
Review Request.

Also we should not try to delete an image that we
do not know anymore. This could only happen if
someone terminates an active push and the Review
Request has an incomplete draft.

+++ contrib/mercurial_git_push.py
@@ -662,7 +662,7 @@ class MercurialReviewRequest(BaseReviewRequest):
                                               only_links='delete')
         hashes = {}
         existing = {}
-        for entry in a:
+        for entry in a.all_items:
             existing[entry['caption']] = entry
 
         def modified(filename):
@@ -695,7 +695,7 @@ class MercurialReviewRequest(BaseReviewRequest):
                 handle_upload(e)
 
         for entry in stored_hashes:
-            if entry not in hashes:
+            if entry not in hashes and entry in existing:
                 existing.get(entry).delete()
 
         return json.dumps(hashes)
